refact: Restrict alloc items behind alloc feat cfg
Because `std` is not included conditionally in this library,
items from the `alloc` crate are only required conditionally,
being replaced with items from the `std` prelude otherwise.

This does create a slight issue, that using `String` as the default type
for `Msg` in the `RJSend` enum, will throw an error
if both `std` and `alloc` are disabled, but the alloc crate is required
via the `serde_json` crate anyway, meaning doing so is erroneous anyway,
and will be covered with a `compile_error!` in future.