feat: Add compile_error! when missing alloc + std
Because using `rjsend` without either the `alloc` or `std` features,
is not valid, it should throw an informative error
when attempting to compile it.
As such, that is what this commit addresses.

This commit does not eliminate other errors
produced by this crate though, notably, due to making `String`
the default type for `Msg` in the `RJSend` enum.
(Consequently, this type may be changed in future...)