docs: Enable automatic feat tagging for docs.rs
Following the decision to introduce conditional items to this crate,
and the intent to release it on `crates.io`, it is useful for `rustdoc`
to be able to add tags to items which are conditionally available
behind features.

As this is currently an unsable feature, it's enabled
by a top-level `cfg_attr` with metadata to enable this `cfg` value
when `rustdoc` is run on `docs.rs`.