fix: Use &'static str for RJSend Msg default type
Because using `String` as the default type for `Msg`
in the `RJSend` enum results in an unknown item error error, when `alloc` is
enabled for `serde_json`, but not `rjsend`, this replaces it with the
`&'static str` type, as an easy shorthand for returning errors
with messages from static strings, without having shackle the type
to add a generic lifetime needed to use non-static strings.

This will result in more boilerplate for apps using the crate
for deserialization, and errors based on dynamic strings,
but it eliminates the error noted above, at minimal realistic cost...

In addition, by removing the use of the `String` type,
it also eliminates the need to import the external `alloc` crate,
and so, this commit also removes the lines to do so.