feat: Add variant unwrapping methods (#12)
As RJSend is effectively a result type, it's useful to provide methods
to provide shorthands for behaviour to use when extracting values
from different variants.

* feat: Add unwrap panic function

The behaviour when unwrapping enum variants,
without otherwise specifying how to handle unexpected variants,
they're generally expected to panic. In the standard library,
this is handled by the `unwrap_failed` function.
This commit adds a similar function to this library,
to handle these cases.

* feat: Add RJSend::unwrap method

As `RJSend` serves a similar purpose to the `Result` enum,
it will be useful to have a methods that assume a particular variant,
and extract the value(s) from it. Consequently, this commit
adds the `RJSend::unwrap` method to extract
the `RJSend::Success` variant.

* feat: Add RJSend::unwrap_fail method

* feat: Add RJSend::unwrap_error method

* feat: Add RJSend::expect method

Similarly to the `unwrap` style methods, `expect` methods,
allowing a custom panic message to be supplied,
can be expected for `RJSend` as well,
due to its similarity to the `Result` enum.
This commit adds the `RJSend::expect` method,
to extract the value from the `RJSend::Success` variant.

* feat: Add RJSend::expect_fail method

* feat: Add RJSend::expect_error method

* feat: Add RJSend::unwrap_or method

As a result type, it's useful to provide infalible methods
for unwrapping the type for unsucessful variants.

Consequently, this commit adds the `RJSend::unwrap_or` method,
mirroring the `Result::unwrap_or` method.

* feat: Add RJSend::unwrap_or_else method

Similarly, to the previous commit, an option
to lazily initialize an alternative value
when unwrapping a result value, which this commit provides,
inline with `Result::unwrap_or_else`.

* feat: Add RJSend::unwrap_or_default method

Similarly to previously, providing a method to use the default value
of a given type as the fallback when unwrapping a result type is useful.
Consequently, this commit adds the `RJSend::unwrap_or_default` method.