fix: Add `RJSend::unwrap_or_default` lint allow
Because Clippy doesn't have the awareness
that `RJSend::unwrap_or_else` is not `Result::unwrap_or_else`
or `Option::unwrap_or_else`, it throws a warning in the implementation
for `RJSend::unwrap_or_default`, suggesting to use `unwrap_or_default`,
which the method is implementing. So this adds an exception
to `RJSend::unwrap_or_default` to silence this lint.