feat: Add option extractor methods (#13)
These commits will add methods similar to
the Result::ok and Result::err methods in the standard library,
for each variant of RJSend, mapping them to an Option
containing its inner value, whilst discarding other variants.

As demonstrated in the standard library, this can be useful
for filtering for a particular variant, when the contents of the others
aren't of interest.

* feat: Add RJSend::success extractor method

* feat: Add RJSend::fail extractor method

* feat: Add RJSend::error extractor method

The extractor method this commit adds, deviates notably
from the format of the other extractors, as the variant it extracts
has several fields. As such, it also acts as a constructor
for an entirely new `ErrorFields` value as well.